CA refactoring
