+++ src/gui/AvailableCourses.java
@@ -1,4 +1,4 @@
-package InputViewDemo;
+package gui;
 
 import java.io.IOException;
 import java.nio.file.Files;
